Revert "WIP. Creating new runner to run compiled star and also adding similar compilation step to isopacks. Need to change ensureIsopacketsLoadable method to recompile crash app everytime in checkout."
This reverts commit 4454f6798989487415606c0155d38f024f3bc539.
Commit is currently broken and I want to get restart removed before putting this in.
